requirements.txt
aiohttp
async
gunicorn
cerbos
promql-parser
prometheus_client>=0.20.0
jwt
opentelemetry-instrumentation-aiohttp-server
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-exporter-zipkin
opentelemetry-distro
opentelemetry-instrumentation-logging
opa-python
